backend/pyproject.toml
[project]
name = "links-backend"
version = "1.0.0"
description = "Social media content ingestion and categorization API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "supabase>=2.0",
    "yt-dlp>=2024.0",
    "instaloader>=4.10",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "google-genai>=1.70.0",
]

[project.scripts]
dev = "uvicorn main:app --reload --port 8000"
